新的敌人完全添加完成(架构构好了加的就是快)

+++ Level.h
@@ -9,6 +9,6 @@ public:
 private:
 	int rand_enemyTime;
 	float rand_enemyPosX;
-	int beforeEnemyPosX;//把前一个生成的敌人坐标记下来，这样可以避免两个敌人互相盖住对方
+	float beforeEnemyPosX;//把前一个生成的敌人坐标记下来，这样可以避免两个敌人互相盖住对方
 	clock_t timeStart, timeEnd;
 };
